requirements.txt
requests>=2.31.0
psycopg[binary]>=3.2.0
sqlglot>=25.0.0
python-dotenv>=1.0.1